# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Marketing OS Lite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brand-profiles:
    get:
      operationId: listBrandProfiles
      tags: [brand-profiles]
      summary: List brand profiles
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of brand profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrandProfile"
    post:
      operationId: createBrandProfile
      tags: [brand-profiles]
      summary: Create a brand profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandProfileBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"
  /brand-profiles/{id}:
    get:
      operationId: getBrandProfile
      tags: [brand-profiles]
      summary: Get a brand profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Brand profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"
    put:
      operationId: updateBrandProfile
      tags: [brand-profiles]
      summary: Update a brand profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandProfileBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandProfile"

  # Campaigns
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List generated assets
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratedAsset"
    post:
      operationId: generateAssets
      tags: [assets]
      summary: Generate content assets for a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAssetsBody"
      responses:
        "201":
          description: Assets generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratedAsset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get a generated asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedAsset"
  /assets/{id}/variants:
    get:
      operationId: getAssetVariants
      tags: [assets]
      summary: Get channel variants for an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Channel variants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelVariant"

  # Approvals
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List approval decisions
      parameters:
        - name: assetId
          in: query
          schema:
            type: integer
        - name: campaignId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of approval decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalDecision"
    post:
      operationId: createApproval
      tags: [approvals]
      summary: Submit an approval decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalBody"
      responses:
        "201":
          description: Approval decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecision"

  # Ad Platform Connections
...
components:
  schemas:
...
    BrandProfile:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        brandName:
          type: string
        toneOfVoice:
          type: string
        targetAudience:
          type: string
        productsServices:
          type: string
        forbiddenClaims:
          type: string
        preferredChannels:
          type: array
          items:
            type: string
        visualNotes:
          type: string
        createdAt:
          type: string
      required: [id, workspaceId, brandName, toneOfVoice, targetAudience, productsServices, forbiddenClaims, preferredChannels, visualNotes, createdAt]
    CreateBrandProfileBody:
      type: object
      properties:
        workspaceId:
          type: integer
        brandName:
          type: string
        toneOfVoice:
          type: string
        targetAudience:
          type: string
        productsServices:
          type: string
        forbiddenClaims:
          type: string
        preferredChannels:
          type: array
          items:
            type: string
        visualNotes:
          type: string
      required: [workspaceId, brandName, toneOfVoice, targetAudience, productsServices, forbiddenClaims, preferredChannels, visualNotes]
...
    GeneratedAsset:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        headline:
          type: string
        shortCaption:
          type: string
        longCaption:
          type: string
        cta:
          type: string
        hashtags:
          type: array
          items:
            type: string
        videoScript:
          type: string
        storyboardOutline:
          type: string
        status:
          type: string
          enum: [draft, reviewed, approved, rejected]
        createdAt:
          type: string
      required: [id, campaignId, headline, shortCaption, longCaption, cta, hashtags, videoScript, storyboardOutline, status, createdAt]
    ChannelVariant:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
        channel:
          type: string
        headline:
          type: string
        caption:
          type: string
        cta:
          type: string
        hashtags:
          type: array
          items:
            type: string
      required: [id, assetId, channel, headline, caption, cta, hashtags]
    GenerateAssetsBody:
      type: object
      properties:
        campaignId:
          type: integer
      required: [campaignId]
    ApprovalDecision:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
        campaignId:
          type: integer
        actor:
          type: string
        decision:
          type: string
          enum: [approved, rejected, changes_requested]
        reason:
          type: string
        createdAt:
          type: string
      required: [id, actor, decision, reason, createdAt]
    CreateApprovalBody:
      type: object
      properties:
        assetId:
          type: integer
        campaignId:
          type: integer
        actor:
          type: string
        decision:
          type: string
          enum: [approved, rejected, changes_requested]
        reason:
          type: string
      required: [actor, decision, reason]
...